+++ main-service/src/main/java/ru/practicum/main/service/compilation/repository/CompilationRepository.java
@@ -8,7 +8,7 @@ import ru.practicum.main.service.compilation.model.Compilation;
 import java.util.List;
 
 @Repository
-public interface CompilationRepository extends JpaRepository<Compilation, Integer> {
+public interface CompilationRepository extends JpaRepository<Compilation, Long> {
 
     List<Compilation> findAllByPinned(Boolean pinned, Pageable pageable);
 }
